Infowindow content layout updated to use the toggle panel view.
