+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-view.js
@@ -33,7 +33,7 @@ module.exports = cdb.core.View.extend({
       label: _t('editor.layers.menu-tab-pane-labels.data'),
       selected: !this.options.selectedNodeId,
       createContentView: function () {
-        return new DataView();
+        return new cdb.core.View();
       }
     }, {
       label: _t('editor.layers.menu-tab-pane-labels.analyses'),
